requirements.txt
massive
python-dotenv
requests
